Refactor: remove DataService.kt - DataService.kt 삭제, 해당 코드가 사용된 부분 리팩토링 진행 완료
