+++ LIBROG/app/src/main/java/com/example/librog/ui/main/addFlowerpot/LockedFlowerpotFragment.kt
@@ -1,43 +1,68 @@
 package com.example.librog.ui.main.addFlowerpot
 
+import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.librog.data.entities.FlowerData
-import com.example.librog.data.remote.data.DataService
+import com.example.librog.ApplicationClass
+import com.example.librog.data.remote.data.DataInterface
+import com.example.librog.data.remote.data.DataResponse3
 import com.example.librog.data.remote.data.LockedFpResult
 import com.example.librog.databinding.FragmentLockedFlowerpotBinding
 import com.example.librog.ui.BaseFragment
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 import kotlin.collections.ArrayList
 
 class LockedFlowerpotFragment :
     BaseFragment<FragmentLockedFlowerpotBinding>(FragmentLockedFlowerpotBinding::inflate) {
 
     private var lockedFpList = ArrayList<LockedFpResult>()
-    private val dataService = DataService
+    private val dataService = ApplicationClass.retrofit.create(DataInterface::class.java)
     private lateinit var adapter: LockedFlowerpotRVAdapter
 
     override fun initAfterBinding() {
+        val userIdx = getUserIdx()
         lockedFpList.clear()
-        dataService.getLockedFpResult(this)
+        getLockedFpResult(userIdx)
         initLockedRv()
+    }
+
+    // 미획득 화분 정보 가져오기
+    private fun getLockedFpResult(userIdx: Int) {
+        dataService.getLockedFpResult(userIdx).enqueue(object : Callback<DataResponse3> {
+            override fun onResponse(call: Call<DataResponse3>, response: Response<DataResponse3>) {
+                val resp = response.body()!!
+                // 실패 시 코드 작성하기
+                when (resp.code) {
+                    1000 -> {
+                        setLockedFpList(resp.result)
+                    }
+                }
+            }
+
+            override fun onFailure(call: Call<DataResponse3>, t: Throwable) {
+            }
+
+        })
     }
 
     private fun initLockedRv() {
         adapter = LockedFlowerpotRVAdapter(lockedFpList)
-        binding.lockedFlowerpotRv.layoutManager =
-            LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
+        binding.lockedFlowerpotRv.layoutManager = LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
         binding.lockedFlowerpotRv.adapter = adapter
-        binding.lockedFlowerpotRv.addItemDecoration(
-            DividerItemDecoration(
-                context,
-                DividerItemDecoration.VERTICAL
-            )
-        )
+        binding.lockedFlowerpotRv.addItemDecoration(DividerItemDecoration(context, DividerItemDecoration.VERTICAL))
     }
 
     fun setLockedFpList(result: ArrayList<LockedFpResult>) {
         lockedFpList.addAll(result)
         adapter.notifyDataSetChanged()
+    }
+
+
+    private fun getUserIdx(): Int{
+        val spf = activity?.getSharedPreferences("userInfo", AppCompatActivity.MODE_PRIVATE)
+        return spf!!.getInt("idx",-1)
     }
 
 }
