+++ LIBROG/app/src/main/java/com/example/librog/ui/main/flowerpot/FlowerpotFragment.kt
@@ -6,42 +6,48 @@ import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.librog.ApplicationClass
 import com.example.librog.R
 import com.example.librog.data.entities.FlowerData
 import com.example.librog.data.entities.Flowerpot
 import com.example.librog.data.local.AppDatabase
-import com.example.librog.data.remote.data.DataService
+import com.example.librog.data.remote.data.DataInterface
+import com.example.librog.data.remote.data.DataResponse1
 import com.example.librog.data.remote.data.FpResult
 import com.example.librog.databinding.FragmentFlowerpotBinding
 import com.example.librog.ui.BaseFragment
 import com.example.librog.ui.main.addFlowerpot.AddFlowerpotActivity
 import com.google.gson.Gson
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 class FlowerpotFragment :
     BaseFragment<FragmentFlowerpotBinding>(FragmentFlowerpotBinding::inflate) {
 
     private var flowerDataList = ArrayList<FlowerData>()
     private var flowerpotList = ArrayList<Flowerpot>()
-    private val dataService = DataService
+    private val dataService = ApplicationClass.retrofit.create(DataInterface::class.java)
     private lateinit var adapter: FlowerpotRVAdapter
-    private lateinit var db: AppDatabase
 
     override fun initAfterBinding() {
+        val userIdx = getUserIdx()
         flowerDataList.clear()
         flowerpotList.clear()
+        getFpList(userIdx)
 
-        db = AppDatabase.getInstance(requireContext())!!
-        dataService.getFpList(this)
 
         adapter = FlowerpotRVAdapter(flowerDataList, flowerpotList)
         binding.flowerpotListRv.adapter = adapter
-        binding.flowerpotListRv.layoutManager =
-            LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
+        binding.flowerpotListRv.layoutManager = LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
         binding.flowerpotListRv.addItemDecoration(DividerItemDecoration(context, DividerItemDecoration.VERTICAL))
+        binding.flowerpotTotalTv.text = String.format(getString(R.string.flowerpot_total), flowerpotList.size)
 
-        binding.flowerpotTotalTv.text =
-            String.format(getString(R.string.flowerpot_total), flowerpotList.size)
+        initClickListener()
 
+    }
+
+    private fun initClickListener(){
         adapter.setMyItemClickListener(object : FlowerpotRVAdapter.OnItemClickListener {
             override fun onItemClick(flowerData: FlowerData, flowerpot: Flowerpot) {
                 val intent = Intent(context, DetailFlowerpotActivity::class.java)
@@ -55,17 +61,47 @@ class FlowerpotFragment :
             }
         })
 
+
         binding.flowerpotAddflowerpotBtn.setOnClickListener {
             val intent = Intent(context, AddFlowerpotActivity::class.java)
             startActivity(intent)
         }
     }
 
-    // getFpList() 실행 시 실행되는 코드
+    //화분 정보 가져오기
+    private fun getFpList(userIdx: Int) {
+        dataService.getFpList(userIdx).enqueue(object : Callback<DataResponse1> {
+            override fun onResponse(call: Call<DataResponse1>, response: Response<DataResponse1>) {
+                val resp = response.body()!!
+                when (resp.code) {
+                    1000 -> {
+                        Log.d("resp", resp.result.size.toString())
+                        setData(resp.result)
+                    }
+
+                    2019 -> {
+
+                    }
+                    2020 -> {
+
+                    }
+                    else -> {
+
+                    }
+                }
+
+            }
+
+            override fun onFailure(call: Call<DataResponse1>, t: Throwable) {
+                t.printStackTrace()
+            }
+        })
+
+    }
+
     // 백엔드 api에서 받은 result 결과를 FlowerData, Flowerpot 형식에 맞게 추가하고 recyclerview에 적용
     fun setData(result: ArrayList<FpResult>){
         for (item in result) {
-
             if (item.flowerDataIdx != null){
                 flowerDataList.add(
                     FlowerData(
@@ -97,18 +133,15 @@ class FlowerpotFragment :
 
                 binding.flowerpotTotalTv.text =
                     String.format(getString(R.string.flowerpot_total), flowerpotList.size)
+
                 adapter.notifyDataSetChanged()
 
             } else {
                 // 유저에 화분이 존재하지 않는 경우
 //                binding.flowerpotListRv.visibility = View.GONE
-
-
-
             }
 
 
-            //이후 비동기 방식으로 room db에 추가할 수 있음
         }
     }
